added trade exit and trade deletion.
- renamed .tradingassistant -> .application subdirectory
- added userDetails to the AuthContext
- added status field to the Trade.js model
- added realizedPnL to the TradeSetting.js Mode ( this is temporarilly added here... )
- added TradeSession.js placeholder.

+++ src/pages/Calculator/components/TradeLogs/index.jsx
@@ -8,7 +8,7 @@ import { v4 as uuidv4 } from 'uuid';
 
 export default function TradeLogs() {
   const [page, setPage] = useState();
-  const [exitTradeID, setExitTradeID] = useState("");
+  const [exitTradeID, setExitTradeID] = useState();
   const [exitModalShown, setExitModalShown] = useState(false);
   
   const [UID, setUID] = useState(uuidv4());//for rerendering.
@@ -19,7 +19,6 @@ export default function TradeLogs() {
       setLogs(logs);
     })
   },[page,UID]);
-  
   
   const logsEl = logs.map(
     /**@param {Trade} log */
@@ -36,14 +35,20 @@ export default function TradeLogs() {
           <td className="text-xs">{Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(Number(log.entryPrice).toFixed(2))}</td>
           <td className="text-xs">{Number(log.leverage).toFixed(2)}</td>
           <td className="text-xs">{new Date(log.createdDate.toDate()).toLocaleDateString()}</td>
-          <td className="text-xs"><span className="badge bg-warning">Open</span></td>
+          <td className="text-xs">
+            <span className={`text-uppercase badge ${log.status === "open" ? "bg-warning" : "bg-success" }`}>
+              {log.status}
+            </span>
+          </td>
           <td className="text-xs">{Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(Number(log.tradeValue).toFixed(2))}</td>
           <td className="text-xs">
-          <div className="btn-group">
+          {log.status === "open" ? 
+            (<div className="btn-group">
             <button type="button" className="btn btn-xs btn-default btn-flat p-1" data-trade-id={log.id}
               onClick={
                 (e)=>{
-                  setExitTradeID(e.currentTarget.getAttribute("data-trade-id"));
+                  const tradeID = e.currentTarget.getAttribute("data-trade-id");
+                  setExitTradeID(tradeID);
                   setExitModalShown(true);
                 }
               }
@@ -51,24 +56,41 @@ export default function TradeLogs() {
               <i className="bi bi-box-arrow-left"></i>
             </button>
             <button type="button" className="btn btn-xs btn-danger btn-flat" data-trade-id={log.id} onClick={(e)=>{
-              const tradeID = e.currentTarget.getAttribute("data-trade-id")
-              tradeService.delete(tradeID).then(function(res){
-                setUID(uuidv4())
-                swal.fire({
-                  title: "Success",
-                  icon: "success",
-                  text: `${tradeID} was successfully deleted.`
-                })
+              const tradeID = e.currentTarget.getAttribute("data-trade-id");
+              swal.fire({
+                title: 'Are you sure?',
+                text: "Trade will be deleted permanently.",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Confirm',
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  tradeService.delete(tradeID).then(function(res){
+                    setUID(uuidv4())
+                    swal.fire({
+                      title: "Success",
+                      icon: "success",
+                      text: `${tradeID} was successfully deleted.`
+                    })
+                  })
+                }else if(result.isDenied){
+                  //action cancelled.
+                }
               })
+              
             }}>
               <i className="bi bi-x-circle"></i>
             </button>
-          </div>
+          </div>) : <p className='text-center m-0 p-0'>-</p>
+          }
           </td>
       </tr>
       );
     }
   )
+  // console.log(exitTradeID);
   logsEl.push(<tr className="p-2"><td colSpan="10" className='text-center text-xs py-2'><a href="javascript:void(0)" onClick={()=>{console.log("Show here was clicked..")}}>Show More...</a></td></tr>)
   return (
     <>
